requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.38
python-dotenv==1.0.0
pydantic>=2.10
httpx==0.25.2
pymysql==1.1.0
cryptography==46.0.3